backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate>=4.0.7

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# AI and ML
crewai>=0.40.0
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0
langchain>=0.1.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data processing
openpyxl>=3.1.0
xlrd>=2.0.0

# API and validation
marshmallow>=3.20.0
werkzeug>=3.0.0

# Production server
gunicorn>=21.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
celery>=5.3.0
redis>=5.0.0

# Logging and monitoring
structlog>=23.2.0

# File handling
python-magic>=0.4.0

# Testing (for development)
pytest>=7.4.0
pytest-cov>=4.1.0
